+++ packages/react/src/button/index.tsx
@@ -1,3 +1,4 @@
-export { Button, ButtonGroup, IconButton } from "@chakra-ui/react";
+export { Button, ButtonGroup } from "@chakra-ui/react";
 export type { ButtonProps, ButtonGroupProps, IconButtonProps } from "@chakra-ui/react";
 // export * from "./Button";
+export * from "./IconButton";
