--- packages/react/src/iconButton/index.tsx
@@ -1,1 +1,0 @@
-export * from "./IconButton";
