+++ packages/react/src/theme/index.ts
@@ -1,6 +1,6 @@
 import { extendTheme } from "@chakra-ui/react";
 import { theme as defaultTheme } from "@chakra-ui/theme";
-// import * as components from "./components";
+import * as components from "./components";
 import * as tokens from "./tokens";
 
 export const theme = extendTheme({
@@ -8,6 +8,6 @@ export const theme = extendTheme({
   ...tokens,
   components: {
     ...defaultTheme.components,
-    // ...components,
+    ...components,
   },
 });
